% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftdoi-package.R
\docType{data}
\name{crossref_member_prefix}
\alias{crossref_member_prefix}
\title{Local cache of metadata on Crossref members}
\format{
A data frame with 576 rows and 5 variables
}
\description{
Local cache of metadata on Crossref members
}
\details{
with columns:
\itemize{
\item id: the Crossref publisher member identifier
\item prefixes: DOI prefixes for the publisher
\item primary_name: the Crossref publisher member name
}
}
\keyword{data}
