% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{ftd_doi}
\alias{ftd_doi}
\title{DOI}
\usage{
ftd_doi(doi, ...)
}
\arguments{
\item{doi}{(character) a DOI. required}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
DOI
}
\examples{
\dontrun{
# found
ftd_doi('10.7554/eLife.07404')

# not found
# ftd_doi('10.1007/1-4020-0613-6_20780')
}
}
