% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{ftd_fetch_patterns}
\alias{ftd_fetch_patterns}
\title{Download patterns files}
\usage{
ftd_fetch_patterns()
}
\value{
character vector of file paths
}
\description{
Does various checks to see if patterns files alrady downloaded,
out of date, if some/all are deleted and in need of an update
}
\examples{
\dontrun{
ftd_fetch_patterns()
}
}
