% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefixes.R
\name{ftd_prefixes}
\alias{ftd_prefixes}
\title{Prefixes}
\usage{
ftd_prefixes(id = NULL, ...)
}
\arguments{
\item{id}{(character) a DOI prefix. Default is \code{NULL}, which
gets all}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Prefixes
}
\examples{
\dontrun{
ftd_prefixes()
ftd_prefixes(id = '10.1080')

# doesn't work
# ftd_prefixes(id = '10.9999')
}
}
