% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi.R
\name{pb_doi}
\alias{pb_doi}
\title{DOI}
\usage{
pb_doi(doi, ...)
}
\arguments{
\item{doi}{(character) a DOI. required}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
DOI
}
\examples{
\dontrun{
# found
pb_doi('10.7554/eLife.07404')

# not found
# pb_doi('10.1007/1-4020-0613-6_20780')
}
}
