% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heartbeat.R
\name{pb_heartbeat}
\alias{pb_heartbeat}
\title{Heartbeat}
\usage{
pb_heartbeat(...)
}
\arguments{
\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Heartbeat
}
\examples{
\dontrun{
pb_heartbeat()
}
}
