% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/members.R
\name{pb_members}
\alias{pb_members}
\title{Members}
\usage{
pb_members(id = NULL, ...)
}
\arguments{
\item{id}{(character) a Crossref member ID. Default is \code{NULL} which
gets all members}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Members
}
\examples{
\dontrun{
df <- pb_members()
pb_members(221)
pb_members(1965)

# not found
# pb_members(999)
}
}
