% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefixes.R
\name{pb_prefixes}
\alias{pb_prefixes}
\title{Prefixes}
\usage{
pb_prefixes(id = NULL, ...)
}
\arguments{
\item{id}{(character) a DOI prefix. Default is \code{NULL}, which
gets all}

\item{...}{curl options passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Prefixes
}
\examples{
\dontrun{
pb_prefixes()
pb_prefixes(id = '10.1080')
pb_prefixes(id = '10.1080', verbose = TRUE)

# doesn't work
# pb_prefixes(id = '10.9999')
}
}
