% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefix_local.R
\name{prefix_local}
\alias{prefix_local}
\title{prefix local}
\usage{
prefix_local(doi)
}
\arguments{
\item{doi}{(characte) a doi}
}
\value{
a named list with: prefix, member, name
}
\description{
prefix local
}
\examples{
prefix_local('10.3390/ani4010082')
}
