% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rftdoi-package.R
\docType{package}
\name{rftdoi-package}
\alias{rftdoi-package}
\alias{rftdoi}
\title{Interface to the ftdoi.org API for publisher url patterns.}
\description{
Interface to the ftdoi.org API for publisher url patterns.
}
\keyword{package}
