% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpubpatterns-package.R
\docType{package}
\name{rpubpatterns-package}
\alias{rpubpatterns-package}
\alias{rpubpatterns}
\title{Interface to the ftdoi.org API for publisher url patterns.}
\description{
Interface to the ftdoi.org API for publisher url patterns.
}
\keyword{package}
